experiment names now propagate to observation_interface
note that this change introduces a third read of the Template file
we may want to be more efficient in the future...
